lint: clippy -D warnings over every target is a gate; ~800 warnings to 0, byte-identical
cargo clippy --workspace --all-targets measured 866 warnings (204 dsp,
215 encode, 30 decode, the rest harnesses/tests). Three moves:

* Per-crate policy allows with their reasons ("Clippy policy" at the top
  of each lib.rs and tests/all/main.rs): the lints that fight a
  line-for-line C port — too_many_arguments and needless_range_loop /
  manual_memcpy (signatures and loop shapes mirror the reference so a
  differential reads against the C), collapsible_if (C control flow with
  line refs between the ifs), manual_clamp (C's chained min/max: `clamp`
  panics on inverted bounds and differs under NaN), manual_range_contains,
  neg_cmp_op_on_partial_ord (the CNN mirrors C float compares incl. NaN),
  excessive_precision (weight tables verbatim), large_enum_variant
  (SbTree), while_let_loop; harness-only: verbatim C constants
  (byte groupings / hex casing), documented identical branches, const-gated
  asserts.
* `cargo clippy --fix` for the machine-applicable rest (94 files), then
  ~70 hand fixes: is_multiple_of, needless borrows and casts, precedence,
  is_some_and, Copy clones, writeln!, const thread_local initializers,
  div_ceil, slice::swap, doc list/quote reflows, unused imports/variables.
* Dead code deleted: cdef_frame_u8_stop, write_intra_y_mode_nonkf,
  av1_inverse_transform_add_u8, ext_tx_set, nstep_radii, zero_cdf_row,
  UV_D67_PRED, a dead per-candidate chroma prediction in the IntraBC
  search (its consumer went in e29e568), and RetainedSlab::len /
  log2_idx / MAX_MVSEARCH_STEPS / DecodeConfig::check_stop / MODE_TO_ANGLE
  kept under cfg(test) (or cfg(feature = "census")) where only tests use
  them.

`just clippy` (harness feature set, -D warnings) is a gate-landing step
and a CI job (it needs the oracle, so it shares the linux leg's cache).

Gates: just gate-landing in a clean worktree with the conformance corpus
linked — upstream-check pristine, fmt-check, clippy, deny, doc-check,
ci-yaml ok; test-next 1548/1548; test-next-scalar 1548/1548 (every byte
gate green after the rewrites); census-gate 4/4; test-whereat 4/4;
api-doc-check green (no public item changed).

+++ crates/aom-bench/tests/all/kb5_lossless_speed_axis.rs
@@ -67,7 +67,7 @@ fn textured_cell(
             // pseudo-noise term — enough coefficient energy that qindex 0 codes
             // real EOBs rather than all-skip.
             let ramp = 32 + (r + col) * 160 / (w + h);
-            let tex = ((r % 3) * 17 + (col % 5) * 11) as usize;
+            let tex = (r % 3) * 17 + (col % 5) * 11;
             let noise = ((r * 2654435761 + col * 40503) >> 7) % 23;
             y[r * w + col] = scale(((ramp + tex + noise) % 256) as u32);
         }
